refactor: extract TUI dashboard rendering from monitor module
Rollback trigger: new module structure causes regressions in UI layout or compilation failures.
