+++ crates/ramshared-cli/src/monitor/mod.rs
@@ -1,23 +1,19 @@
 //! Read-only RamShared observability stream and terminal dashboard.
 
-use std::collections::{BTreeMap, VecDeque};
+use std::collections::BTreeMap;
 use std::fmt;
 use std::fs::{self, OpenOptions};
 use std::io::Write;
 use std::path::{Path, PathBuf};
 use std::process::Command;
-use std::time::{Duration, Instant, SystemTime, UNIX_EPOCH};
+use std::time::{Duration, SystemTime, UNIX_EPOCH};
 
-use ratatui::crossterm::event::{self, Event, KeyCode, KeyEventKind, KeyModifiers};
-use ratatui::layout::{Constraint, Direction, Layout, Rect};
-use ratatui::style::{Color, Style};
-use ratatui::text::Line;
-use ratatui::widgets::{Block, Borders, Paragraph, Sparkline, Wrap};
-use ratatui::{DefaultTerminal, Frame};
 use serde::{Deserialize, Serialize};
 use serde_json::{Map, Value};
 
-use crate::{bounded_process, cascade, workload};
+use crate::{bounded_process, cascade};
+
+pub mod tui_dashboard;
 
 const DEFAULT_INTERVAL_MS: u64 = 1_000;
 const DEFAULT_HISTORY_SECONDS: u64 = 300;
@@ -515,7 +511,7 @@ fn count_scope_dirs(path: &Path) -> u64 {
 }
 
 fn read_reservation_totals(path: &Path) -> (u64, u64) {
-    workload::read_reservation_ledger(path).map_or_else(
+    crate::workload::read_reservation_ledger(path).map_or_else(
         |_| (0, 0),
         |reservations| {
             let reserved_bytes = reservations
@@ -748,7 +744,7 @@ pub fn run(options: &MonitorOptions) -> Result<(), MonitorError> {
     if options.jsonl || options.once {
         return run_jsonl(options);
     }
-    run_tui(options)
+    tui_dashboard::run_tui(options)
 }
 
 fn run_compact(options: &MonitorOptions) -> Result<(), MonitorError> {
@@ -840,44 +836,6 @@ fn write_atomic(path: &Path, line: &str) -> Result<(), MonitorError> {
     fs::rename(&temporary, path).map_err(|error| MonitorError::Io(error.to_string()))
 }
 
-#[derive(Default)]
-struct TierAccumulator {
-    min_mbs: f64,
-    max_mbs: f64,
-    total_mbs: f64,
-    count: u64,
-}
-
-impl TierAccumulator {
-    fn record(&mut self, speed: f64) {
-        if speed >= 5.0 {
-            self.min_mbs = if self.min_mbs == 0.0 {
-                speed
-            } else {
-                self.min_mbs.min(speed)
-            };
-            self.max_mbs = self.max_mbs.max(speed);
-            self.total_mbs += speed;
-            self.count += 1;
-        }
-    }
-
-    fn avg_mbs(&self) -> f64 {
-        if self.count > 0 {
-            self.total_mbs / self.count as f64
-        } else {
-            0.0
-        }
-    }
-
-    fn apply_to_plane_io(&self, plane_io: &mut TierIoStats) {
-        plane_io.min_mbs = self.min_mbs;
-        plane_io.avg_mbs = self.avg_mbs();
-        plane_io.max_mbs = self.max_mbs;
-        plane_io.peak_mbs = self.max_mbs;
-    }
-}
-
 fn update_tier_latencies(
     cp: &mut ControlPlaneObservation,
     zram_count: u64,
@@ -921,345 +879,9 @@ fn update_tier_latencies(
     };
 }
 
-fn should_exit_tui(event_opt: Option<Event>) -> bool {
-    if let Some(Event::Key(key)) = event_opt {
-        key.kind == KeyEventKind::Press
-            && (matches!(key.code, KeyCode::Char('q') | KeyCode::Esc)
-                || (key.code == KeyCode::Char('c')
-                    && key.modifiers.contains(KeyModifiers::CONTROL)))
-    } else {
-        false
-    }
-}
-
-fn run_tui(options: &MonitorOptions) -> Result<(), MonitorError> {
-    let mut terminal = ratatui::init();
-    let result = tui_loop(&mut terminal, options);
-    ratatui::restore();
-    result
-}
-
-fn tui_loop(terminal: &mut DefaultTerminal, options: &MonitorOptions) -> Result<(), MonitorError> {
-    let history_limit =
-        ((options.history_seconds * 1_000) / options.interval_ms).clamp(1, 10_000) as usize;
-    let mut history = VecDeque::with_capacity(history_limit);
-    let interval = Duration::from_millis(options.interval_ms);
-    let mut next_sample = Instant::now();
-    let mut observation = collect_observation()?;
-    let mut last_io_sample = Some((
-        observation.control_plane.swap_read_bytes,
-        observation.control_plane.swap_write_bytes,
-        observation.control_plane.zram_io.read_bytes,
-        observation.control_plane.zram_io.write_bytes,
-        observation.control_plane.vram_io.read_bytes,
-        observation.control_plane.vram_io.write_bytes,
-        observation.control_plane.disk_io.read_bytes,
-        observation.control_plane.disk_io.write_bytes,
-        Instant::now(),
-    ));
-    let mut last_faults_sample = Some((
-        observation.control_plane.pgfault_total,
-        observation.control_plane.pgmajfault_total,
-    ));
-
-    let mut zram_acc = TierAccumulator::default();
-    let mut vram_acc = TierAccumulator::default();
-    let mut disk_acc = TierAccumulator::default();
-    let mut swap_peak_mbs = 0.0f64;
-    let mut zram_peak_used_mb = 0u64;
-    let mut vram_peak_used_mb = 0u64;
-    let mut disk_peak_used_mb = 0u64;
-
-    loop {
-        if Instant::now() >= next_sample {
-            if let Ok(new_obs) = collect_observation() {
-                observation = new_obs;
-            }
-            let now = Instant::now();
-            if let Some((
-                last_rb,
-                last_wb,
-                last_z_rb,
-                last_z_wb,
-                last_v_rb,
-                last_v_wb,
-                last_d_rb,
-                last_d_wb,
-                last_t,
-            )) = last_io_sample
-            {
-                let dt = now.duration_since(last_t).as_secs_f64();
-                if (0.05..=10.0).contains(&dt) {
-                    let divisor = dt * 1_048_576.0;
-                    let cp = &mut observation.control_plane;
-                    cp.swap_read_mbs =
-                        (cp.swap_read_bytes.saturating_sub(last_rb) as f64) / divisor;
-                    cp.swap_write_mbs =
-                        (cp.swap_write_bytes.saturating_sub(last_wb) as f64) / divisor;
-
-                    cp.zram_io.read_mbs =
-                        (cp.zram_io.read_bytes.saturating_sub(last_z_rb) as f64) / divisor;
-                    cp.zram_io.write_mbs =
-                        (cp.zram_io.write_bytes.saturating_sub(last_z_wb) as f64) / divisor;
-
-                    cp.vram_io.read_mbs =
-                        (cp.vram_io.read_bytes.saturating_sub(last_v_rb) as f64) / divisor;
-                    cp.vram_io.write_mbs =
-                        (cp.vram_io.write_bytes.saturating_sub(last_v_wb) as f64) / divisor;
-
-                    cp.disk_io.read_mbs =
-                        (cp.disk_io.read_bytes.saturating_sub(last_d_rb) as f64) / divisor;
-                    cp.disk_io.write_mbs =
-                        (cp.disk_io.write_bytes.saturating_sub(last_d_wb) as f64) / divisor;
-
-                    let total_swap_speed = cp.swap_read_mbs + cp.swap_write_mbs;
-                    swap_peak_mbs = swap_peak_mbs.max(total_swap_speed);
-                    cp.swap_peak_mbs = swap_peak_mbs;
-
-                    zram_acc.record(cp.zram_io.read_mbs + cp.zram_io.write_mbs);
-                    zram_acc.apply_to_plane_io(&mut cp.zram_io);
-
-                    vram_acc.record(cp.vram_io.read_mbs + cp.vram_io.write_mbs);
-                    vram_acc.apply_to_plane_io(&mut cp.vram_io);
-
-                    disk_acc.record(cp.disk_io.read_mbs + cp.disk_io.write_mbs);
-                    disk_acc.apply_to_plane_io(&mut cp.disk_io);
-
-                    if let Some((last_pf, last_mpf)) = last_faults_sample {
-                        cp.pgfault_per_sec =
-                            (cp.pgfault_total.saturating_sub(last_pf) as f64 / dt) as u64;
-                        cp.pgmajfault_per_sec =
-                            (cp.pgmajfault_total.saturating_sub(last_mpf) as f64 / dt) as u64;
-                    }
-                }
-            }
-
-            if let Some(tiers) = observation.value("tiers").and_then(Value::as_object) {
-                if let Some(t) = tiers.get("zram").and_then(Value::as_object) {
-                    let u = t.get("used_kib").and_then(Value::as_u64).unwrap_or(0);
-                    zram_peak_used_mb = zram_peak_used_mb.max((u + 512) / 1024);
-                }
-                if let Some(t) = tiers.get("vram").and_then(Value::as_object) {
-                    let u = t.get("used_kib").and_then(Value::as_u64).unwrap_or(0);
-                    vram_peak_used_mb = vram_peak_used_mb.max((u + 512) / 1024);
-                }
-                if let Some(t) = tiers.get("disk").and_then(Value::as_object) {
-                    let u = t.get("used_kib").and_then(Value::as_u64).unwrap_or(0);
-                    disk_peak_used_mb = disk_peak_used_mb.max((u + 512) / 1024);
-                }
-            }
-
-            let cp = &mut observation.control_plane;
-            cp.swap_peak_mbs = swap_peak_mbs;
-            cp.zram_peak_used_mb = zram_peak_used_mb;
-            cp.vram_peak_used_mb = vram_peak_used_mb;
-            cp.disk_peak_used_mb = disk_peak_used_mb;
-            zram_acc.apply_to_plane_io(&mut cp.zram_io);
-            vram_acc.apply_to_plane_io(&mut cp.vram_io);
-            disk_acc.apply_to_plane_io(&mut cp.disk_io);
-
-            update_tier_latencies(cp, zram_acc.count, vram_acc.count, disk_acc.count);
-
-            last_io_sample = Some((
-                cp.swap_read_bytes,
-                cp.swap_write_bytes,
-                cp.zram_io.read_bytes,
-                cp.zram_io.write_bytes,
-                cp.vram_io.read_bytes,
-                cp.vram_io.write_bytes,
-                cp.disk_io.read_bytes,
-                cp.disk_io.write_bytes,
-                now,
-            ));
-            last_faults_sample = Some((cp.pgfault_total, cp.pgmajfault_total));
-            if let Ok(flight_line) = serde_json::to_string(&observation) {
-                let _ = fs::write("/dev/shm/ramshared-flight.json", format!("{flight_line}\n"));
-            }
-            history.push_back(memory_used_pct(&observation.mem));
-            while history.len() > history_limit {
-                history.pop_front();
-            }
-            next_sample = Instant::now() + interval;
-        }
-        let _ = terminal.draw(|frame| draw_dashboard(frame, &observation, &history));
-
-        let wait = next_sample
-            .saturating_duration_since(Instant::now())
-            .min(Duration::from_millis(100));
-        let event_opt = if event::poll(wait).unwrap_or(false) {
-            event::read().ok()
-        } else {
-            None
-        };
-        if should_exit_tui(event_opt) {
-            return Ok(());
-        }
-    }
-}
-
-fn memory_used_pct(memory: &MemoryObservation) -> u64 {
-    if memory.total_kib == 0 {
-        return 0;
-    }
-    memory
-        .total_kib
-        .saturating_sub(memory.available_kib)
-        .saturating_mul(100)
-        / memory.total_kib
-}
-
-fn draw_dashboard(frame: &mut Frame<'_>, observation: &Observation, history: &VecDeque<u64>) {
-    let rows = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([
-            Constraint::Length(3),
-            Constraint::Percentage(35),
-            Constraint::Percentage(55),
-            Constraint::Length(2),
-        ])
-        .split(frame.area());
-    let top = Layout::default()
-        .direction(Direction::Horizontal)
-        .constraints([Constraint::Percentage(50), Constraint::Percentage(50)])
-        .split(rows[1]);
-    let bottom = Layout::default()
-        .direction(Direction::Horizontal)
-        .constraints([Constraint::Percentage(60), Constraint::Percentage(40)])
-        .split(rows[2]);
-
-    let daemon_alive = observation
-        .value("daemon")
-        .and_then(|d| d.get("alive"))
-        .and_then(Value::as_bool)
-        .unwrap_or(false);
-
-    let (status_text, state_color) = if daemon_alive {
-        ("🟢 STATUS: OPERATIONAL & PROTECTED", Color::Green)
-    } else if observation.bool_value("ok") == Some(true) {
-        ("🟢 STATUS: OPERATIONAL", Color::Green)
-    } else {
-        ("🟡 STATUS: ARMED & READY", Color::Yellow)
-    };
-
-    let version = env!("CARGO_PKG_VERSION");
-    let header = Paragraph::new(Line::from(format!(
-        " RamShared v{version} │ {status_text} │ Phase: {} │ Memory Protection: ACTIVE",
-        observation.string("phase"),
-    )))
-    .style(Style::default().fg(state_color))
-    .block(
-        Block::default()
-            .borders(Borders::ALL)
-            .title("System Overview"),
-    );
-    frame.render_widget(header, rows[0]);
-
-    draw_memory(frame, top[0], observation, history);
-    draw_gpu(frame, top[1], observation);
-    draw_tiers(frame, bottom[0], observation);
-    draw_control(frame, bottom[1], observation);
-    frame.render_widget(
-        Paragraph::new(" [q / Esc]: exit │ Priority Order: RAM (1st) -> GPU VRAM (2nd) -> Host SSD (3rd) (highest filled first)"),
-        rows[3],
-    );
-}
-
-fn draw_memory(
-    frame: &mut Frame<'_>,
-    area: Rect,
-    observation: &Observation,
-    history: &VecDeque<u64>,
-) {
-    let chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([Constraint::Length(5), Constraint::Min(1)])
-        .split(area);
-    let memory = &observation.mem;
-    let total_mb = (memory.total_kib + 512) / 1024;
-    let avail_mb = (memory.available_kib + 512) / 1024;
-    let used_mb = total_mb.saturating_sub(avail_mb);
-    let used_pct = (used_mb * 100).checked_div(total_mb).unwrap_or(0);
-
-    let bar_len: u64 = 20;
-    let filled = (used_pct * bar_len / 100).min(bar_len);
-    let empty = bar_len.saturating_sub(filled);
-    let bar = format!(
-        "[{}{}]",
-        "█".repeat(filled as usize),
-        "░".repeat(empty as usize)
-    );
-
-    let swap_used = (memory.swap_total_kib.saturating_sub(memory.swap_free_kib) + 512) / 1024;
-    let swap_total = (memory.swap_total_kib + 512) / 1024;
-    let swap_pct = (swap_used * 100).checked_div(swap_total).unwrap_or(0);
-    let swap_bar = make_bar(swap_pct, bar_len);
-    let text = format!(
-        " Host RAM:  {bar} {used_pct:>2}% ({used_mb:>5} MB / {total_mb} MB)\n Total Swap: {swap_bar} {swap_pct:>2}% ({swap_used:>5} MB / {swap_total} MB)\n Pressure:   Light Stall (Some): {psi_some:.2}% │ Severe Stall (Full): {psi_full:.2}%",
-        psi_some = observation.control_plane.memory_psi_some_avg10,
-        psi_full = observation.control_plane.memory_psi_full_avg10,
-    );
-    frame.render_widget(
-        Paragraph::new(text).block(
-            Block::default()
-                .borders(Borders::ALL)
-                .title("Host RAM & Swap"),
-        ),
-        chunks[0],
-    );
-    let values: Vec<u64> = history.iter().copied().collect();
-    frame.render_widget(
-        Sparkline::default()
-            .block(Block::default().borders(Borders::ALL).title("RAM History"))
-            .data(&values)
-            .max(100),
-        chunks[1],
-    );
-}
-
-fn draw_gpu(frame: &mut Frame<'_>, area: Rect, observation: &Observation) {
-    let text = observation.gpu.as_ref().map_or_else(
-        || " GPU not detected".to_string(),
-        |gpu| {
-            let used_pct = gpu
-                .used_mib
-                .saturating_mul(100)
-                .checked_div(gpu.total_mib)
-                .unwrap_or(0);
-            let bar_len: u64 = 20;
-            let filled = (used_pct.saturating_mul(bar_len) / 100).min(bar_len);
-            let empty = bar_len.saturating_sub(filled);
-            let bar = format!(
-                "[{}{}]",
-                "█".repeat(filled as usize),
-                "░".repeat(empty as usize)
-            );
-            format!(
-                " Graphics Card: {}\n GPU VRAM:      {bar} {used_pct:>2}% ({} MB / {} MB)\n Available VRAM: {} MB free\n PCIe Hardware:  PCIe Gen 3 x16 │ Bandwidth: 8.74 GB/s (8,950 MB/s)",
-                gpu.name, gpu.used_mib, gpu.total_mib, gpu.free_mib
-            )
-        },
-    );
-    frame.render_widget(
-        Paragraph::new(text)
-            .block(Block::default().borders(Borders::ALL).title("GPU"))
-            .wrap(Wrap { trim: true }),
-        area,
-    );
-}
-
-fn make_bar(pct: u64, len: u64) -> String {
-    let filled = (pct.saturating_mul(len) / 100).min(len);
-    let empty = len.saturating_sub(filled);
-    format!(
-        "[{}{}]",
-        "█".repeat(filled as usize),
-        "░".repeat(empty as usize)
-    )
-}
-
 fn make_tier_bar(used: u64, size: u64, len: u64) -> String {
     if size == 0 {
-        return make_bar(0, len);
+        return tui_dashboard::make_bar(0, len);
     }
     let mut filled = (used.saturating_mul(len) / size).min(len);
     if filled == 0 && used > 0 {
@@ -1273,409 +895,12 @@ fn make_tier_bar(used: u64, size: u64, len: u64) -> String {
     )
 }
 
-fn compute_tier_speedup(io: &TierIoStats, tier_prio: i32) -> String {
-    let ssd_baseline = 20.0f64;
-    match tier_prio {
-        100 => {
-            if io.max_mbs >= 5.0 {
-                let min_mult = (io.min_mbs / ssd_baseline).clamp(1.0, 500.0);
-                let avg_mult = (io.avg_mbs / ssd_baseline).clamp(1.0, 500.0);
-                let max_mult = (io.max_mbs / ssd_baseline).clamp(1.0, 500.0);
-                format!(
-                    "⚡ Min: {:.0}x │ Avg: {:.0}x │ Max: {:.0}x (Active vs Host VHDX)",
-                    min_mult, avg_mult, max_mult
-                )
-            } else {
-                "⚡ 250x In-RAM Capable (0.05 µs)".to_string()
-            }
-        }
-        50 => {
-            if io.max_mbs >= 5.0 {
-                let min_mult = (io.min_mbs / ssd_baseline).clamp(1.0, 150.0);
-                let avg_mult = (io.avg_mbs / ssd_baseline).clamp(1.0, 150.0);
-                let max_mult = (io.max_mbs / ssd_baseline).clamp(1.0, 150.0);
-                format!(
-                    "🚀 Min: {:.0}x │ Avg: {:.0}x │ Max: {:.0}x (Active vs Host VHDX)",
-                    min_mult, avg_mult, max_mult
-                )
-            } else {
-                "🚀 20x-100x PCIe DMA Capable (8.74 GB/s)".to_string()
-            }
-        }
-        _ => {
-            if io.max_mbs >= 5.0 {
-                "🐢 Min: 1.0x │ Avg: 1.0x │ Max: 1.0x (WSL2 System Disk)".to_string()
-            } else {
-                "🐢 1.0x Host VHDX Baseline (WSL2 System Disk)".to_string()
-            }
-        }
-    }
-}
-
-fn format_tier_latency(
-    io: &TierIoStats,
-    default_min: f64,
-    default_avg: f64,
-    default_max: f64,
-    suffix: &str,
-) -> String {
-    let min = if io.min_lat_us > 0.0 {
-        io.min_lat_us
-    } else {
-        default_min
-    };
-    let avg = if io.avg_lat_us > 0.0 {
-        io.avg_lat_us
-    } else {
-        default_avg
-    };
-    let max = if io.max_lat_us > 0.0 {
-        io.max_lat_us
-    } else {
-        default_max
-    };
-
-    if max >= 1000.0 {
-        format!("{min:.0}..{avg:.0}..{:.1}ms ({suffix})", max / 1000.0)
-    } else {
-        format!("{min:.2}..{avg:.2}..{max:.2}µs ({suffix})")
-    }
-}
-
-fn draw_tiers(frame: &mut Frame<'_>, area: Rect, observation: &Observation) {
-    let width = area.width;
-    let bar_len = ((width as u64) / 8).clamp(8, 20);
-    let sep_len = (width.saturating_sub(4) as usize).max(20);
-    let sep = "─".repeat(sep_len);
-
-    let tiers = observation
-        .value("tiers")
-        .and_then(Value::as_object)
-        .map(|tiers| {
-            // Helper to extract tier data with proper rounding
-            let get = |name: &str| -> (bool, u64, u64) {
-                let tier = tiers.get(name).and_then(Value::as_object);
-                let present = tier
-                    .and_then(|t| t.get("present"))
-                    .and_then(Value::as_bool)
-                    .unwrap_or(false);
-                let used_kib = tier
-                    .and_then(|t| t.get("used_kib"))
-                    .and_then(Value::as_u64)
-                    .unwrap_or(0);
-                let size_kib = tier
-                    .and_then(|t| t.get("size_kib"))
-                    .and_then(Value::as_u64)
-                    .unwrap_or(0);
-                let used = (used_kib + 512) / 1024;
-                let size = (size_kib + 512) / 1024;
-                (present, used, size)
-            };
-
-            let (zram_on, zram_used, zram_size) = get("zram");
-            let (vram_on, vram_used, vram_size) = get("vram");
-            let (disk_on, disk_used, disk_size) = get("disk");
-
-            let zram_status = if !zram_on {
-                "⚫ OFF"
-            } else if zram_used > 0 {
-                "🟢 ACTIVE [1st Target]"
-            } else {
-                "🟢 ARMED [1st Target]"
-            };
-
-            let vram_status = if !vram_on {
-                "⚫ OFF"
-            } else if vram_used > 0 {
-                "🟢 ACTIVE [2nd Target]"
-            } else {
-                "🟢 ARMED [2nd Target]"
-            };
-
-            let disk_status = if !disk_on {
-                "⚫ OFF"
-            } else if disk_used > 0 {
-                "🔵 COLD BOOT BASELINE"
-            } else {
-                "🔵 STANDBY [3rd Fallback]"
-            };
-
-            let zram_pct = zram_used
-                .saturating_mul(100)
-                .checked_div(zram_size)
-                .unwrap_or(0);
-            let vram_pct = vram_used
-                .saturating_mul(100)
-                .checked_div(vram_size)
-                .unwrap_or(0);
-            let disk_pct = disk_used
-                .saturating_mul(100)
-                .checked_div(disk_size)
-                .unwrap_or(0);
-
-            let z_r = observation.control_plane.zram_io.read_mbs;
-            let z_w = observation.control_plane.zram_io.write_mbs;
-            let v_r = observation.control_plane.vram_io.read_mbs;
-            let v_w = observation.control_plane.vram_io.write_mbs;
-            let d_r = observation.control_plane.disk_io.read_mbs;
-            let d_w = observation.control_plane.disk_io.write_mbs;
-
-            let z_min = observation.control_plane.zram_io.min_mbs;
-            let z_avg = observation.control_plane.zram_io.avg_mbs;
-            let z_max = observation.control_plane.zram_io.max_mbs;
-
-            let v_min = observation.control_plane.vram_io.min_mbs;
-            let v_avg = observation.control_plane.vram_io.avg_mbs;
-            let v_max = observation.control_plane.vram_io.max_mbs;
-
-            let d_min = observation.control_plane.disk_io.min_mbs;
-            let d_avg = observation.control_plane.disk_io.avg_mbs;
-            let d_max = observation.control_plane.disk_io.max_mbs;
-
-            let zram_speedup = compute_tier_speedup(&observation.control_plane.zram_io, 100);
-            let vram_speedup = compute_tier_speedup(&observation.control_plane.vram_io, 50);
-            let disk_speedup = compute_tier_speedup(&observation.control_plane.disk_io, -2);
-
-            let z_lat = format_tier_latency(
-                &observation.control_plane.zram_io,
-                0.04,
-                0.08,
-                0.15,
-                "In-RAM LZ4",
-            );
-            let v_lat = format_tier_latency(
-                &observation.control_plane.vram_io,
-                0.85,
-                1.45,
-                3.20,
-                "PCIe DMA",
-            );
-            let d_lat = format_tier_latency(
-                &observation.control_plane.disk_io,
-                85.0,
-                180.0,
-                1200.0,
-                "Host VHDX",
-            );
-
-            let z_bar = make_tier_bar(zram_used, zram_size, bar_len);
-            let v_bar = make_tier_bar(vram_used, vram_size, bar_len);
-            let d_bar = make_tier_bar(disk_used, disk_size, bar_len);
-
-            let z_pct_str = if zram_used > 0 && zram_pct == 0 {
-                "<1%".to_string()
-            } else {
-                format!("{zram_pct:>2}%")
-            };
-            let v_pct_str = if vram_used > 0 && vram_pct == 0 {
-                "<1%".to_string()
-            } else {
-                format!("{vram_pct:>2}%")
-            };
-            let d_pct_str = if disk_used > 0 && disk_pct == 0 {
-                "<1%".to_string()
-            } else {
-                format!("{disk_pct:>2}%")
-            };
-
-            let z_peak = observation.control_plane.zram_peak_used_mb.max(zram_used);
-            let v_peak = observation.control_plane.vram_peak_used_mb.max(vram_used);
-            let d_peak = observation.control_plane.disk_peak_used_mb.max(disk_used);
-
-            let z_peak_pct = z_peak
-                .saturating_mul(100)
-                .checked_div(zram_size)
-                .unwrap_or(0);
-            let v_peak_pct = v_peak
-                .saturating_mul(100)
-                .checked_div(vram_size)
-                .unwrap_or(0);
-            let d_peak_pct = d_peak
-                .saturating_mul(100)
-                .checked_div(disk_size)
-                .unwrap_or(0);
-
-            let z_use = format!(
-                "{z_bar} {z_pct_str} ( {zram_u:>4} MB / {zram_t} MB ) │ Peak: {z_peak:>4} MB ({z_peak_pct:>3}%)",
-                z_bar = z_bar,
-                z_pct_str = z_pct_str,
-                zram_u = zram_used,
-                zram_t = zram_size,
-                z_peak = z_peak,
-                z_peak_pct = z_peak_pct
-            );
-            let v_use = format!(
-                "{v_bar} {v_pct_str} ( {vram_u:>4} MB / {vram_t} MB ) │ Peak: {v_peak:>4} MB ({v_peak_pct:>3}%)",
-                v_bar = v_bar,
-                v_pct_str = v_pct_str,
-                vram_u = vram_used,
-                vram_t = vram_size,
-                v_peak = v_peak,
-                v_peak_pct = v_peak_pct
-            );
-            let d_use = format!(
-                "{d_bar} {d_pct_str} ( {disk_u:>4} MB / {disk_t} MB ) │ Peak: {d_peak:>4} MB ({d_peak_pct:>3}%)",
-                d_bar = d_bar,
-                d_pct_str = d_pct_str,
-                disk_u = disk_used,
-                disk_t = disk_size,
-                d_peak = d_peak,
-                d_peak_pct = d_peak_pct
-            );
-
-            let format_rate = |min: f64, avg: f64, max: f64| {
-                if max >= 5.0 {
-                    format!("Min: {min:>4.0} │ Avg: {avg:>4.0} │ Max: {max:>4.0} MB/s")
-                } else {
-                    "Idle (Awaiting Workload)".to_string()
-                }
-            };
-            let z_rate = format_rate(z_min, z_avg, z_max);
-            let v_rate = format_rate(v_min, v_avg, v_max);
-            let d_rate = format_rate(d_min, d_avg, d_max);
-
-            format!(
-                concat!(
-                    " ╔══ 📦 TIER 1: RAM Swap (zram) ── Priority: 100 ── {zram_s}\n",
-                    " ║   ├─ Memory Usage:       {z_use}\n",
-                    " ║   ├─ Real-Time Speed:    Read: {z_r:>5.1} MB/s │ Write: {z_w:>5.1} MB/s\n",
-                    " ║   ├─ Throughput Stats:   {z_rate}\n",
-                    " ║   ├─ Hardware Latency:   {z_lat}\n",
-                    " ║   └─ Speedup Factor:     {zram_speedup}\n",
-                    " ╠{sep}\n",
-                    " ║   🚀 TIER 2: GPU VRAM (nbd0) ── Priority:  50 ── {vram_s}\n",
-                    " ║   ├─ Memory Usage:       {v_use}\n",
-                    " ║   ├─ Real-Time Speed:    Read: {v_r:>5.1} MB/s │ Write: {v_w:>5.1} MB/s\n",
-                    " ║   ├─ Throughput Stats:   {v_rate}\n",
-                    " ║   ├─ Hardware Latency:   {v_lat}\n",
-                    " ║   └─ Speedup Factor:     {vram_speedup}\n",
-                    " ╠{sep}\n",
-                    " ║   💾 TIER 3: WSL2 System Disk ── Priority:  -2 ── {disk_s}\n",
-                    " ║   ├─ Memory Usage:       {d_use}\n",
-                    " ║   ├─ Real-Time Speed:    Read: {d_r:>5.1} MB/s │ Write: {d_w:>5.1} MB/s\n",
-                    " ║   ├─ Throughput Stats:   {d_rate}\n",
-                    " ║   ├─ Hardware Latency:   {d_lat}\n",
-                    " ║   └─ Speedup Factor:     {disk_speedup}\n",
-                    " ╚{sep}",
-                ),
-                zram_s = zram_status,
-                z_use = z_use,
-                z_r = z_r,
-                z_w = z_w,
-                z_rate = z_rate,
-                z_lat = z_lat,
-                zram_speedup = zram_speedup,
-                sep = sep,
-                vram_s = vram_status,
-                v_use = v_use,
-                v_r = v_r,
-                v_w = v_w,
-                v_rate = v_rate,
-                v_lat = v_lat,
-                vram_speedup = vram_speedup,
-                disk_s = disk_status,
-                d_use = d_use,
-                d_r = d_r,
-                d_w = d_w,
-                d_rate = d_rate,
-                d_lat = d_lat,
-                disk_speedup = disk_speedup,
-            )
-        })
-        .unwrap_or_else(|| " Swap Tiers: not available".to_string());
-
-    frame.render_widget(
-        Paragraph::new(tiers).block(
-            Block::default()
-                .borders(Borders::ALL)
-                .title("Memory Tiers (Swap Priority & Speedup)"),
-        ),
-        area,
-    );
-}
-
-fn draw_control(frame: &mut Frame<'_>, area: Rect, observation: &Observation) {
-    let width = area.width;
-    let sep_len = (width.saturating_sub(4) as usize).max(20);
-    let sep = "─".repeat(sep_len);
-
-    let pid = observation
-        .value("daemon")
-        .and_then(|d| d.get("pid"))
-        .and_then(Value::as_u64)
-        .map(|p| p.to_string())
-        .unwrap_or_else(|| "-".to_string());
-
-    let daemon_alive = observation
-        .value("daemon")
-        .and_then(|d| d.get("alive"))
-        .and_then(Value::as_bool)
-        .unwrap_or(false);
-
-    let errors = if observation.errors.is_empty() {
-        "None".to_string()
-    } else {
-        format!("{}", observation.errors.len())
-    };
-
-    let swap_in = observation.control_plane.swap_in_pages;
-    let swap_out = observation.control_plane.swap_out_pages;
-    let read_mbs = observation.control_plane.swap_read_mbs;
-    let write_mbs = observation.control_plane.swap_write_mbs;
-    let peak_mbs = observation.control_plane.swap_peak_mbs;
-    let pgfault_rate = observation.control_plane.pgfault_per_sec;
-    let pgmajfault_rate = observation.control_plane.pgmajfault_per_sec;
-    let boot_info = match observation.control_plane.boot_tier_latency_ms {
-        Some(ms) => format!("{:.2}s (Tier Ready)", ms as f64 / 1000.0),
-        None => "3.12s (Tier Ready)".to_string(),
-    };
-
-    let text = format!(
-        concat!(
-            " Daemon Status:            {daemon_icon} {daemon_txt} (PID {pid})\n",
-            " Boot Initialization:      ⏱️  {boot_info}\n",
-            " Safety Guard:             🛡️  Fail-Closed (Zero Panic)\n",
-            " Swap I/O Protocol:        ⚡ Synchronous Zero-Copy (.rw_page)\n",
-            " PCIe Hardware Link:       🚀 Gen 3 x16 (8.74 GB/s DMA)\n",
-            " {sep}\n",
-            " Real-Time Speed:          Read: {read_mbs:>4.1} MB/s │ Write: {write_mbs:>4.1} MB/s\n",
-            " Peak Recorded Speed:      🚀 {peak_mbs:>5.1} MB/s (Latching Max)\n",
-            " Cumulative Page I/O:      In: {swap_in} pages │ Out: {swap_out} pages\n",
-            " Page Faults Rate:         📊 {pgfault_rate}/s (Major: {pgmajfault_rate}/s)\n",
-            " Anomaly Counter:          {errors}\n",
-            " {sep}\n",
-            " ⚡ ALLOCATION GUARANTEE:\n",
-            " All new writes fill RAM (1st) and VRAM (2nd) before SSD.\n",
-            " SSD usage is cold WSL2 boot baseline.",
-        ),
-        daemon_icon = if daemon_alive { "🟢" } else { "🔴" },
-        daemon_txt = if daemon_alive { "RUNNING" } else { "STOPPED" },
-        pid = pid,
-        boot_info = boot_info,
-        read_mbs = read_mbs,
-        write_mbs = write_mbs,
-        peak_mbs = peak_mbs,
-        swap_in = swap_in,
-        swap_out = swap_out,
-        pgfault_rate = pgfault_rate,
-        pgmajfault_rate = pgmajfault_rate,
-        errors = errors,
-        sep = sep,
-    );
-
-    frame.render_widget(
-        Paragraph::new(text).block(
-            Block::default()
-                .borders(Borders::ALL)
-                .title("Diagnostics & Live Stats"),
-        ),
-        area,
-    );
-}
-
 #[cfg(test)]
-mod tests {
+pub mod tests {
     #![allow(clippy::expect_used, clippy::unwrap_used)]
+    use std::collections::VecDeque;
+    use std::time::Instant;
+    use crate::monitor::tui_dashboard::*;
 
     use super::*;
     use crate::workload;
@@ -1683,7 +908,7 @@ mod tests {
     use ratatui::backend::TestBackend;
     use std::os::unix::fs::PermissionsExt;
 
-    fn observation(ok: bool, with_gpu: bool) -> Observation {
+    pub fn observation(ok: bool, with_gpu: bool) -> Observation {
         let status = serde_json::from_value::<BTreeMap<String, Value>>(serde_json::json!({
             "schema_version": 4,
             "ok": ok,
@@ -1757,18 +982,29 @@ mod tests {
     }
 
     fn monitor_ledger_path(name: &str, contents: Option<&str>) -> (PathBuf, PathBuf) {
-        let root = std::env::temp_dir().join(format!(
-            "ramshared-monitor-ledger-{name}-{}",
-            std::process::id()
-        ));
-        let _ = fs::remove_dir_all(&root);
-        fs::create_dir_all(&root).unwrap();
-        let path = root.join("reservations.json");
-        if let Some(contents) = contents {
-            fs::write(&path, contents).unwrap();
-        }
-        (root, path)
+    let root = std::env::temp_dir().join(format!(
+        "ramshared-monitor-ledger-{name}-{}",
+        std::process::id()
+    ));
+    let _ = fs::remove_dir_all(&root);
+    fs::create_dir_all(&root).unwrap();
+
+    // Set directory permissions
+    let mut root_perms = fs::metadata(&root).unwrap().permissions();
+    use std::os::unix::fs::PermissionsExt;
+    root_perms.set_mode(0o700);
+    fs::set_permissions(&root, root_perms).unwrap();
+
+    let path = root.join("reservations.json");
+    if let Some(contents) = contents {
+        fs::write(&path, contents).unwrap();
+        // Set file permissions
+        let mut file_perms = fs::metadata(&path).unwrap().permissions();
+        file_perms.set_mode(0o600);
+        fs::set_permissions(&path, file_perms).unwrap();
     }
+    (root, path)
+}
 
     #[test]
     // TestName: monitor_telemetry_refuses_missing_reservation_ledger
@@ -2081,113 +1317,6 @@ mod tests {
 
         assert!(started.elapsed() < Duration::from_millis(750));
         assert!(error.contains("output"), "{error}");
-    }
-
-    #[test]
-    fn computes_dynamic_tier_speedup_values() {
-        let idle_io = TierIoStats::default();
-        assert_eq!(
-            compute_tier_speedup(&idle_io, 100),
-            "⚡ 250x In-RAM Capable (0.05 µs)"
-        );
-        assert_eq!(
-            compute_tier_speedup(&idle_io, 50),
-            "🚀 20x-100x PCIe DMA Capable (8.74 GB/s)"
-        );
-        assert_eq!(
-            compute_tier_speedup(&idle_io, -2),
-            "🐢 1.0x Host VHDX Baseline (WSL2 System Disk)"
-        );
-
-        let zram_active = TierIoStats {
-            min_mbs: 100.0,
-            avg_mbs: 460.0,
-            max_mbs: 860.0,
-            ..TierIoStats::default()
-        };
-        let z_txt = compute_tier_speedup(&zram_active, 100);
-        assert!(
-            z_txt.contains("Min: 5x") && z_txt.contains("Avg: 23x") && z_txt.contains("Max: 43x")
-        );
-
-        let vram_active = TierIoStats {
-            min_mbs: 40.0,
-            avg_mbs: 200.0,
-            max_mbs: 600.0,
-            ..TierIoStats::default()
-        };
-        let v_txt = compute_tier_speedup(&vram_active, 50);
-        assert!(
-            v_txt.contains("Min: 2x") && v_txt.contains("Avg: 10x") && v_txt.contains("Max: 30x")
-        );
-
-        let disk_active = TierIoStats {
-            min_mbs: 15.0,
-            avg_mbs: 85.0,
-            max_mbs: 120.0,
-            ..TierIoStats::default()
-        };
-        let d_txt = compute_tier_speedup(&disk_active, -2);
-        assert!(d_txt.contains("Min: 1.0x"));
-    }
-
-    #[test]
-    fn dashboard_renders_cleanly_across_multiple_terminal_resolutions() {
-        let mut sample = observation(true, true);
-        sample.errors = vec!["gpu_dropped".to_string()];
-        sample.control_plane.boot_tier_latency_ms = Some(2890);
-        let history = VecDeque::from(vec![25, 30, 45, 60, 55]);
-        let resolutions = [(80, 24), (100, 30), (140, 40), (200, 50), (240, 60)];
-
-        for (w, h) in resolutions {
-            let backend = ratatui::backend::TestBackend::new(w, h);
-            let mut terminal = ratatui::Terminal::new(backend).unwrap();
-            terminal
-                .draw(|frame| draw_dashboard(frame, &sample, &history))
-                .unwrap();
-            let rendered = terminal
-                .backend()
-                .buffer()
-                .content()
-                .iter()
-                .map(|cell| cell.symbol())
-                .collect::<String>();
-            assert!(
-                rendered.contains("Host RAM") || rendered.contains("RAM"),
-                "Failed at {w}x{h}"
-            );
-            assert!(
-                rendered.contains("Memory Tiers") || rendered.contains("Swap"),
-                "Failed at {w}x{h}"
-            );
-        }
-    }
-
-    #[test]
-    fn edge_case_draw_and_pct_helpers() {
-        assert_eq!(memory_used_pct(&MemoryObservation::default()), 0);
-        assert_eq!(make_bar(0, 10), "[░░░░░░░░░░]");
-        assert_eq!(make_bar(100, 10), "[██████████]");
-
-        let mut sample_no_gpu = observation(false, false);
-        sample_no_gpu.gpu = None;
-        sample_no_gpu.status.remove("tiers");
-
-        let backend = ratatui::backend::TestBackend::new(120, 30);
-        let mut terminal = ratatui::Terminal::new(backend).unwrap();
-        let history = VecDeque::new();
-        terminal
-            .draw(|frame| draw_dashboard(frame, &sample_no_gpu, &history))
-            .unwrap();
-        let rendered = terminal
-            .backend()
-            .buffer()
-            .content()
-            .iter()
-            .map(|cell| cell.symbol())
-            .collect::<String>();
-        assert!(rendered.contains("GPU not detected"));
-        assert!(rendered.contains("Swap Tiers: not available"));
     }
 
     #[test]
